+++ Scripts/Runtime/Core/Rig/AddTeleportToRaycastMask.cs
@@ -1,13 +1,13 @@
 using UnityEngine;
 using UnityEngine.XR.Interaction.Toolkit;
 
-[RequireComponent(typeof(XRRayInteractor))]
+[RequireComponent(typeof(UnityEngine.XR.Interaction.Toolkit.Interactors.XRRayInteractor))]
 public class AddTeleportToRaycastMask : MonoBehaviour
 {
     // Start is called before the first frame update
     void Start()
     {
-        var interactor = GetComponent<XRRayInteractor>();
+        var interactor = GetComponent<UnityEngine.XR.Interaction.Toolkit.Interactors.XRRayInteractor>();
         if (interactor.interactionLayers == (interactor.interactionLayers | (1 << InteractionLayerMask.NameToLayer("Teleport"))))
         {
             interactor.raycastMask |= 1 << LayerMask.NameToLayer("Teleport");