+++ Scripts/Runtime/Core/Rig/HandInteractorManager.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+public class HandInteractorManager : MonoBehaviour
+{
+    [SerializeField] InteractorControl[] m_interactorControls;
+
+    void Update()
+    {
+        foreach (var control in m_interactorControls)
+            control.interactor.SetActive(control.IsActive);
+    }
+}