+++ Scripts/Runtime/Core/Rig/InteractorControl.cs
@@ -1,0 +1,11 @@
+using System;
+using UnityEngine;
+using UnityEngine.XR.Interaction.Toolkit.Inputs.Readers;
+
+[Serializable]
+public class InteractorControl
+{
+    public GameObject interactor;
+    public XRInputButtonReader controlAction;
+    public bool IsActive => controlAction.ReadIsPerformed();
+}