+++ Scripts/Runtime/Core/Rig/RigOrchestrator.cs
@@ -26,28 +26,10 @@ public class RigOrchestrator : MonoBehaviour
 
     public void SwitchToRay()
     {
-        if (TryGetInteractorManager<HandManager>(out var handManager))
-            handManager.InteractionMode = InteractionMode.Ray;
-        else
-            Debug.Log("HandManager not found.");
-
         if (TryGetInteractorManager<HandMenuManager>(out var handMenuManager))
             handMenuManager.Blocked = true;
         else
             Debug.LogWarning("No HandMenuManager for the right hand found.");
-    }
-
-    public void SwitchToTeleport()
-    {
-        if (TryGetInteractorManager<HandManager>(out var handManager))
-            handManager.InteractionMode = InteractionMode.Teleport;
-        else
-            Debug.Log("HandManager not found.");
-
-        if (TryGetInteractorManager<HandMenuManager>(out var handMenuManager))
-            handMenuManager.Blocked = false;
-        else
-            Debug.LogWarning("No HandMenuManager found.");
     }
 
     public void ToggleHandMenu(bool value)
@@ -55,34 +37,10 @@ public class RigOrchestrator : MonoBehaviour
         if (!TryGetInteractorManager<HandMenuManager>(out var handMenuManager) ||
             (handMenuManager != null && handMenuManager.Blocked)) return;
 
-        if (TryGetInteractorManager<HandManager>(out var handManager))
-            handManager.ToggleHandMenu(value);
-        else
-            Debug.Log("HandManager not found.");
-
         if (TryGetInteractorManager<GazeManager>(out var gazeManager))
             gazeManager.BlockInteractor(value);
         else
             Debug.LogWarning("No GazeManager found.");
-    }
-
-    public void BlockTeleport(bool value)
-    {
-        if (TryGetInteractorManager<HandManager>(out var handManager))
-            handManager.TeleportBlocked = value;
-        else
-            Debug.LogWarning("No HandManager found.");
-    }
-    
-    public void BlockInteraction(bool value)
-    {
-        if (TryGetInteractorManager<HandManager>(out var handManager))
-        {
-            handManager.LeftInteractionBlocked = value;
-            handManager.RightInteractionBlocked = value;
-        }
-        else
-            Debug.LogWarning("No HandManager found.");
     }
     
     public bool TryGetInteractorManager<T>(out T interactorManager) where T : InteractorManager