+++ Scripts/Runtime/Misc/ToggleGameObject.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public class ToggleGameObject : MonoBehaviour
+{
+    public void ToggleObject(GameObject go) => go.SetActive(!go.activeSelf);
+}